Let's see what this does

+++ .tests/cloud-init_test.bash
@@ -6,14 +6,11 @@
 
 set -Cefuxo pipefail
 
-path=$(realpath "$0"); init="$(dirname $path)/../cloud-init/init.bash"
+path=$(realpath "$0"); init="$(dirname $path)/../cloud-init/init.yaml"
 
-cpus='1' mem='4GB' disk='32GB'
+name='transcraft' cpus='1' mem='4GB' disk='32GB'
 
-name='transcraft'
-exec=('lsb_release' '-a')
-
-multipass launch --name "$name" -c$cpus -m$mem -d$disk --cloud-init "$init"
-multipass delete --purge foo
-
-multipass exec "$name" -- "${exec[@]}"
+if multipass launch --name "$name" -c$cpus -m$mem -d$disk --cloud-init "$init"
+then
+  multipass delete --purge foo
+fi
